SSCS-8346 Schedule2 optional and test refactor
