+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/WriteFinalDecisionMidEventValidationHandlerTestBase.java
@@ -2,14 +2,13 @@ package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.openMocks;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.MID_EVENT;
 
 import java.time.LocalDate;
-import java.util.Iterator;
 import javax.validation.Validation;
+import javax.validation.Validator;
 import junitparams.JUnitParamsRunner;
 import junitparams.NamedParameters;
 import junitparams.Parameters;
@@ -23,18 +22,7 @@ import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Appellant;
-import uk.gov.hmcts.reform.sscs.ccd.domain.BenefitType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.CaseDetails;
-import uk.gov.hmcts.reform.sscs.ccd.domain.DirectionType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.DocumentLink;
-import uk.gov.hmcts.reform.sscs.ccd.domain.DynamicList;
-import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Identity;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Name;
-import uk.gov.hmcts.reform.sscs.ccd.domain.RegionalProcessingCenter;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+import uk.gov.hmcts.reform.sscs.ccd.domain.*;
 import uk.gov.hmcts.reform.sscs.service.DecisionNoticeQuestionService;
 import uk.gov.hmcts.reform.sscs.service.DecisionNoticeService;
 
@@ -42,7 +30,7 @@ import uk.gov.hmcts.reform.sscs.service.DecisionNoticeService;
 public abstract class WriteFinalDecisionMidEventValidationHandlerTestBase {
 
     protected static final String USER_AUTHORISATION = "Bearer token";
-    protected WriteFinalDecisionMidEventValidationHandler handler;
+    protected static WriteFinalDecisionMidEventValidationHandler handler;
 
     protected abstract String getBenefitType();
 
@@ -68,10 +56,13 @@ public abstract class WriteFinalDecisionMidEventValidationHandlerTestBase {
 
     protected abstract void setValidPointsAndActivitiesScenario(SscsCaseData caseData, String descriptorFlowValue);
 
+    protected static Validator validator = Validation.buildDefaultValidatorFactory().getValidator();
+
     @Before
     public void setUp() {
         openMocks(this);
-        handler = new WriteFinalDecisionMidEventValidationHandler(Validation.buildDefaultValidatorFactory().getValidator(), decisionNoticeService);
+
+        handler = new WriteFinalDecisionMidEventValidationHandler(validator, decisionNoticeService);
 
         when(decisionNoticeService.getQuestionService(getBenefitType())).thenReturn(decisionNoticeQuestionService);
 
@@ -262,35 +253,7 @@ public abstract class WriteFinalDecisionMidEventValidationHandlerTestBase {
         assertEquals("You need to upload PDF documents only", error);
     }
 
-    @Test
-    @Parameters({
-        "STANDARD_RATE, STANDARD_RATE",
-        "ENHANCED_RATE, ENHANCED_RATE",
-        "NOT_CONSIDERED, STANDARD_RATE",
-        "STANDARD_RATE, NO_AWARD",
-        "NO_AWARD, ENHANCED_RATE"
-    })
-    public void shouldExhibitBenefitSpecificBehaviourWhenAnAnAwardIsGivenAndNoActivitiesSelected(AwardType dailyLiving, AwardType mobility) {
-
-        setValidPointsAndActivitiesScenario(sscsCaseData, "Yes");
-        setEmptyActivitiesListScenario(sscsCaseData);
-
-        sscsCaseData.setPipWriteFinalDecisionDailyLivingQuestion(dailyLiving.getKey());
-        sscsCaseData.setPipWriteFinalDecisionMobilityQuestion(mobility.getKey());
-
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(MID_EVENT, callback, USER_AUTHORISATION);
-        assertEquals(0, response.getWarnings().size());
-
-        if ("PIP".equals(getBenefitType())) {
-            assertEquals(1, response.getErrors().size());
-            String error = response.getErrors().stream().findFirst().orElse("");
-            assertEquals("At least one activity must be selected unless there is no award", error);
-        } else {
-            assertEquals(1, response.getErrors().size());
-            String error = response.getErrors().stream().findFirst().orElse("");
-            assertEquals("At least one activity must be selected.", error);
-        }
-    }
+    protected abstract void shouldExhibitBenefitSpecificBehaviourWhenAnAnAwardIsGivenAndNoActivitiesSelected(AwardType dailyLiving, AwardType mobility);
 
     protected abstract void setNoAwardsScenario(SscsCaseData caseData);
 
@@ -298,27 +261,11 @@ public abstract class WriteFinalDecisionMidEventValidationHandlerTestBase {
 
     protected abstract void setNullActivitiesListScenario(SscsCaseData caseData);
 
+    protected abstract void shouldExhibitBenefitSpecificBehaviourWhenNoAwardsAreGivenAndNoActivitiesAreSelected();
 
-    @Test
-    public void shouldExhibitBenefitSpecificBehaviourWhenNoAwardsAreGivenAndNoActivitiesAreSelected() {
+    protected abstract void shouldExhibitBenefitSpecificBehaviourWhenNoAwardsAreGivenAndNoActivitiesAreSelectedAndEndDateTypeIsSetEndDate();
 
-        setValidPointsAndActivitiesScenario(sscsCaseData, "Yes");
-        setNoAwardsScenario(sscsCaseData);
-        setEmptyActivitiesListScenario(sscsCaseData);
-
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(MID_EVENT, callback, USER_AUTHORISATION);
-
-        assertEquals(0, response.getWarnings().size());
-        if ("PIP".equals(getBenefitType())) {
-            assertEquals(0, response.getErrors().size());
-            assertEquals("na", caseDetails.getCaseData().getWriteFinalDecisionEndDateType());
-        } else {
-            assertEquals(1, response.getErrors().size());
-            String error = response.getErrors().stream().findFirst().orElse("");
-            assertEquals("At least one activity must be selected.", error);
-            assertNull(caseDetails.getCaseData().getWriteFinalDecisionEndDateType());
-        }
-    }
+    protected abstract void shouldExhibitBenefitSpecificBehaviourWhenNoAwardsAreGivenAndNoActivitiesAreSelectedAndEndDateTypeIsIndefinite();
 
     @Test
     public void shouldExhibitBenefitSpecificBehaviourWhenNoAwardsAreGivenAndNoActivitiesAreSelectedAndEndDateTypeIsNA() {
@@ -331,90 +278,11 @@ public abstract class WriteFinalDecisionMidEventValidationHandlerTestBase {
         sscsCaseData.setWriteFinalDecisionEndDateType("na");
         PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(MID_EVENT, callback, USER_AUTHORISATION);
 
-        Iterator<String> iterator = response.getErrors().iterator();
-        while (iterator.hasNext()) {
-            System.out.println(iterator.next());
-        }
-
         assertEquals(0, response.getWarnings().size());
 
-        if ("PIP".equals(getBenefitType())) {
-            assertEquals(0, response.getErrors().size());
-        } else {
-            assertEquals(1, response.getErrors().size());
-            String error = response.getErrors().stream().findFirst().orElse("");
-            assertEquals("At least one activity must be selected.", error);
-        }
+        assertEquals(0, response.getErrors().size());
 
         assertEquals("na", caseDetails.getCaseData().getWriteFinalDecisionEndDateType());
-
-    }
-
-    @Test
-    public void shouldExhibitBenefitSpecificBehaviourWhenNoAwardsAreGivenAndNoActivitiesAreSelectedAndEndDateTypeIsSetEndDate() {
-
-        setValidPointsAndActivitiesScenario(sscsCaseData, "Yes");
-        setNoAwardsScenario(sscsCaseData);
-        setEmptyActivitiesListScenario(sscsCaseData);
-
-        sscsCaseData.setWriteFinalDecisionEndDateType("setEndDate");
-
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(MID_EVENT, callback, USER_AUTHORISATION);
-
-        if ("PIP".equals(getBenefitType())) {
-            assertEquals(1, response.getErrors().size());
-            assertEquals("End date is not applicable for this decision - please specify 'N/A - No Award'.", response.getErrors().iterator().next());
-            assertEquals(0, response.getWarnings().size());
-        } else {
-            assertEquals(1, response.getErrors().size());
-            assertEquals("At least one activity must be selected.", response.getErrors().iterator().next());
-            assertEquals(0, response.getWarnings().size());
-        }
-        assertEquals("setEndDate", caseDetails.getCaseData().getWriteFinalDecisionEndDateType());
-
-    }
-
-    @Test
-    public void shouldDisplayActivitiesErrorForPipOnlyWhenNoAwardsAreGivenAndNoActivitiesAreSelectedAndEndDateTypeIsIndefinite() {
-
-        setValidPointsAndActivitiesScenario(sscsCaseData, "Yes");
-        setEmptyActivitiesListScenario(sscsCaseData);
-
-        sscsCaseData.setPipWriteFinalDecisionDailyLivingQuestion(AwardType.NO_AWARD.getKey());
-        sscsCaseData.setPipWriteFinalDecisionMobilityQuestion(AwardType.NO_AWARD.getKey());
-        sscsCaseData.setWriteFinalDecisionEndDateType("indefinite");
-
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(MID_EVENT, callback, USER_AUTHORISATION);
-
-        assertEquals(0, response.getWarnings().size());
-        if ("PIP".equals(getBenefitType())) {
-            assertEquals(1, response.getErrors().size());
-            assertEquals("End date is not applicable for this decision - please specify 'N/A - No Award'.", response.getErrors().iterator().next());
-        } else {
-            assertEquals(1, response.getErrors().size());
-            String error = response.getErrors().stream().findFirst().orElse("");
-            assertEquals("At least one activity must be selected.", error);
-        }
-
-        assertEquals("indefinite", caseDetails.getCaseData().getWriteFinalDecisionEndDateType());
-
-    }
-
-    @Test
-    public void shouldNotDisplayActivitiesErrorWhenActivitiesAreNotYetSelected() {
-
-        setValidPointsAndActivitiesScenario(sscsCaseData, "Yes");
-
-        sscsCaseData.setPipWriteFinalDecisionDailyLivingQuestion(AwardType.STANDARD_RATE.getKey());
-        sscsCaseData.setPipWriteFinalDecisionMobilityQuestion(AwardType.STANDARD_RATE.getKey());
-        sscsCaseData.setPipWriteFinalDecisionDailyLivingActivitiesQuestion(null);
-        sscsCaseData.setPipWriteFinalDecisionMobilityActivitiesQuestion(null);
-
-        PreSubmitCallbackResponse<SscsCaseData> response = handler.handle(MID_EVENT, callback, USER_AUTHORISATION);
-        assertEquals(0, response.getErrors().size());
-        assertEquals(0, response.getWarnings().size());
-
-        assertNull(caseDetails.getCaseData().getWriteFinalDecisionEndDateType());
 
     }
